✨ feat(prisma): extension with specific models to apply soft-delete and filter
