+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "moderate-nestjs",
-    "version": "0.0.13",
+    "version": "0.0.14",
     "description": "REST API Backend using NestJS, Prisma, PostgreSQL, and Docker",
     "author": "hubts",
     "private": true,
