+++ src/infrastructure/prisma/type/prisma-tx-client.type.ts
@@ -1,4 +1,4 @@
-import { ExtendedPrismaClient } from "./extended-prisma-client";
+import { ExtendedPrismaClient } from "../extended-prisma-client";
 
 export type PrismaTxClient = Parameters<
     Parameters<ExtendedPrismaClient["$transaction"]>[0]
